ADD: Minimal required version check (for mainline releases)
Problem:
The original test version selection logic works on assumption that tested feature/fix is present in all
Firebird versions higher than one specifid in test. It's not true for all backported code, as we usually
backport to latest maintenance version. Result is that test could be selected to run against Firebird
version that doesn't support it and test will not pass, generating false alarm.

Solution:
A new test attribute 'min_versions' was introduced. If value is present, it must contain a Firebird version
or list of versions separated by semicolon. The new test version selection logic checks whether the tested
Firebird version isn't from the same "mainline" (major.minor version part) as any listed one, and if it is,
then tests isn't run if tested Firebird versions is not equal or higher than specified one. For example:

min_versions = '2.0.3;2.1.1'
test is run for Firebird versions 1.x, 2.0.3 and higher 2.0.x, 2.1.1 and higher 2.1.x, and mainline higher
than 2.1. Of course, test is not run if no suitable test version is found.